Updated the tests for the discrete canopy code and fixed a few bugs
