+++ eradiate/scenes/biosphere/__init__.py
@@ -1,11 +1,6 @@
 from ._core import BiosphereFactory, Canopy
-from ._discrete import (
-    AbstractTree,
-    CanopyElement,
-    DiscreteCanopy,
-    InstancedCanopyElement,
-    LeafCloud,
-)
+from ._discrete import CanopyElement, DiscreteCanopy, InstancedCanopyElement, LeafCloud
+from ._tree import AbstractTree, MeshTree
 
 __all__ = [
     "BiosphereFactory",
