+++ eradiate/scenes/biosphere/_tree.py
